tracker: sort abbreviations by last updated
When editing abbreviations, you want to be able to see what the last
edited abbreviation is. Because tracked abbreviations used to be sorted
by creation time, if the list of abbreviations got long and truncated,
you could only see the first created abbreviations but not the last
edited. Abbreviations are now sorted by last updated to fix that.